Removed fetch.ts
MessageRequest (previously Request) will stay in a separate file for tidiness
